Fix Remove Items from Cart. Add TradeQuery. A User's previous orders, if this User lends, borrows, sells now determine if he/she is allowed to continue

+++ phase2/src/main/java/phase2/trade/itemlist/TradeItemHolder.java
@@ -27,11 +27,11 @@ public class TradeItemHolder extends ItemList {
         this.setOfItems = items;
     }
 
-    public long getLendCount() {
-        return setOfItems.stream().filter(item -> item.getWillingness().equals(Willingness.LEND)).count();
+    public int getLendCount() {
+        return Math.toIntExact(setOfItems.stream().filter(item -> item.getWillingness().equals(Willingness.LEND)).count());
     }
 
-    public long getSellCount() {
-        return setOfItems.stream().filter(item -> item.getWillingness().equals(Willingness.SELL)).count();
+    public int getSellCount() {
+        return Math.toIntExact(setOfItems.stream().filter(item -> item.getWillingness().equals(Willingness.SELL)).count());
     }
 }
